requirements.txt
torch==1.8.0
torchvision
click
gdown
psutil
Pillow
scipy
h5py
tensorboard
requests
tqdm
pyspng
ninja
imageio-ffmpeg==0.4.3
torchinfo
sigpy
hydra-core==1.0.6
opencv-python
matplotlib
scikit-image==0.19.3
numba==0.52.0
numpy==1.20.0
monai